Fix build with latest ARM Gcc toolchain (11.3.rel1)
Seems we were relying on undefined behaviour when replacing _malloc_r
et-al with our own thread-safe variants in newlib-stubs.c. The latest
toolchain complains of duplicate symbols; the previous did not.

While meddling with this stuff, implement the retargetable locks for
newlib.

Use the linker's "--wrap" feature to override both _malloc_r (et-al) and
the locking primitives. For some reason the default locking primitives
do not have the 'weak' attribute on their symbols, otherwise "--wrap"
would not be required.